Advertise and require xbosh stream restarts.
This is strict about receiving restarts, because with a non-proxy BOSH
server, failing to restart when needed or restarting when not expected
would desynchronize the BOSH server from the XMPP server.  We can detect
this reliably more easily than a proxy can, so do so.

The spec doesn't explain what should happen if an empty request is
received when a restart is needed.  This can easily happen if a pipelining
client sends empty requests in the background while authentication is
taking place.  Stream restarts should be handled like pause and terminate
requests: trigger immediately, flushing any earlier requests in the
pipeline, which is what's implemented here.