Treat 'wait' as per-session, not per-request.
Instead of treating 'wait' as the length of time each request can go
unanswered, treat it as the amount of time the session can go without
responding to at least one held request.  There's no point in responding
to every held request after the wait period; all this does is flush
the entire wait queue, increasing latency if we need to send a response
before the new held request comes in, and increasing traffic.

The spec is vague about this, but this is the only behavior that makes
sense.