Fix rerequests.
If RID#1 received a response and RID#2 was held awaiting one, and new
requests are received for both RID#1 and RID#2, the response to RID#1
needs to be resent, and RID#2 needs to go back into outbound_requests
awaiting a response.

Previously, RID#1 would receive its response, but RID#2 would trigger
an error, since it didn't yet have a response.